chore: rename `PartnerMetrics` related schemas and service functions
